Moved fetching functionality to mainpage, added pull to refresh functionality to businessList component
