+++ Actions/isLoading.js
@@ -1,6 +1,6 @@
 export const IS_LOADING = 'is_loading';
 
-export function isLoading() {
+export function setLoading() {
   return {
     type: IS_LOADING,
   };
